fix check.

+++ cdc/sinkv2/eventsink/mq/worker.go
@@ -335,7 +335,7 @@ func (w *worker) sendMessages(ctx context.Context) error {
 					return err
 				}
 			}
-			w.metricMQWorkerFlushDuration.Observe(float64(time.Since(start).Seconds()))
+			w.metricMQWorkerFlushDuration.Observe(time.Since(start).Seconds())
 		}
 	}
 }
